requirements.txt
requests
beautifulsoup4
streamlit
selenium
openpyxl
Pillow
selenium
webdriver-manager
